% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{constrainSizeFinImage}
\alias{constrainSizeFinImage}
\title{constrainSizeFinImage}
\usage{
constrainSizeFinImage(fin)
}
\arguments{
\item{fin}{Value of type cimg. Load the image via load.image("directory/finImage.JPG")}
}
\value{
Value of type list containing a scalar resize factor and a resized cimg
}
\details{
Processes an image(cimg) containing a fin. 
constrains image to a size range that balances preserving detail and efficiency
}

