% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blobber.R
\name{cropFins}
\alias{cropFins}
\title{cropFins}
\usage{
cropFins(imageName, cropNet, workingImage, saveDir, minXY = 100,
  threshold = 0.45)
}
\arguments{
\item{imageName}{path to image to be cropped}

\item{cropNet}{mxnet model for isolating dolphin}

\item{workingImage}{workspace for holding image during processing}

\item{saveDir}{directory to save cropped image}

\item{minXY}{minimum width/height in pixels for valid crop}
}
\details{
Function loads image and uses the given neural network to isolate any dorsal fins therin.
The neural network returns a matrix whose values range [0,1].
Using a contiguous region algorithm (blob analysis),
the x,y range for each excitation on the matrix is found.
The range is then rescaled to find the analogous region in the original image.
This region is cropped from said image and saved as a jpeg to the given directory.
}
