% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathEncoding.R
\name{decodePath}
\alias{decodePath}
\title{coordinateEncoding}
\usage{
decodePath(coordinateEncoding)
}
\arguments{
\item{coordinateEncoding}{rle object with start coordinates}
}
\value{
matrix of x,y pixel coordinates
}
\description{
Function to decompress rle + start coordinates into image pixel coordinates for trailing edge
}

