% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{distanceToRef}
\alias{distanceToRef}
\title{distanceToRef}
\usage{
distanceToRef(queryHash, referenceHash)
}
\arguments{
\item{queryHashData}{vector containing a hash for matching}

\item{referenceHashData}{matrix containing a reference catalogue of hashes}
}
\description{
Function to calculate the distances from a single query hash to a reference catalogue
}

