% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathEncoding.R
\name{encodePath}
\alias{encodePath}
\title{coordinateMatrix}
\usage{
encodePath(coordinateMatrix)
}
\arguments{
\item{coordinateMatrix}{vector containing a hash for matching}
}
\value{
A vector represenatation of an rle object with start ing pixel coordinates
}
\description{
Function to compress pixel coordinates for trailing edge
}

