% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extractAnnulus}
\alias{extractAnnulus}
\title{extractAnnulus}
\usage{
extractAnnulus(imageFromR, xCoordinates, yCoordinates)
}
\arguments{
\item{imageFromR}{cimg wraped as NumericVector}

\item{xCoordinates}{x pixel coordinates for data extraction}

\item{yCoordinates}{y pixel coordinates for data extraction}
}
\description{
This function returns a cimg wrapped as NumeriMatrix.
Contains simplified image reduced to relevant identifying features.
Features represent 16 point ring from around each coordinate
The x component in the returned cimg represents each coordinate
The y component from the returned cimg represents each sample from the 16 point ring
}
