% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fillGlare}
\alias{fillGlare}
\title{fillGlare}
\usage{
fillGlare(imageFromR, highlightCoordinates, aveThresh = 7L)
}
\arguments{
\item{imageFromR}{cimg wraped as NumericVector}

\item{highlightCoordinates}{Coordinates to fill in}

\item{aveThresh}{Weighted sum, representing minimum number of neighboring non 0 pixels}
}
\description{
This function returns a cimg wrapped as NumeriMatrix.
Using a smoothing kernel, highlights are filled in based on the neighboring values.
Iteratively fills in from outer edge inward.
Image should have glare cliped to 0 already
}

