% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findPath}
\alias{findPath}
\title{findPath}
\usage{
findPath(edge, startx, starty, endx, endy, minX, maxX, minY, maxY, proximity)
}
\arguments{
\item{edge}{Weighted canny edge matrix}

\item{startx}{Starting pixel x component}

\item{starty}{Starting pixel y component}

\item{endx}{Ending pixel x component}

\item{endy}{Ending pixel y component}

\item{minX}{Left boundary}

\item{maxX}{Right boundary}

\item{minY}{Upper boundary}

\item{maxY}{Lower boundary}

\item{proximity}{Distance to ending coordinate at which trace can terminate}
}
\description{
This function returns a NumeriMatrix, identifying the optimal edge
Using the A* algorithm on a canny edge matrix weighted by a neural network
The traversal balances clarity/smoothness of the edge against the network evaluation
}
