% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\docType{class}
\name{finIter-class}
\alias{finIter}
\alias{finIter-class}
\title{An S4 class to wrap mx.io.arrayiter for use in finFindR}
\description{
initialize generates an mx.io.arrayiter
value called from mxnet predict method to process data
}
\section{Methods}{

\describe{
\item{\code{iter.next()}}{bool iter.next()}

\item{\code{num.pad()}}{int num.pad()  const }

\item{\code{reset()}}{void reset()}

\item{\code{value()}}{Rcpp::List value()  const }
}}

