% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocpuApi.R
\name{hashFromImage}
\alias{hashFromImage}
\title{hashFromImage}
\usage{
hashFromImage(imageobj)
}
\arguments{
\item{imageobj}{Value of type cimg. Load the image via load.image("directory/finImage.JPG")}
}
\value{
Value of type list containing:
"coordinates" a matrix of coordinates
"hash" vector specifying an individual
}
\details{
\code{traceFromImage} wrapper for use through opencpu.
opencpu passes temp object name to function followed by \code{traceToHash}
curl -v http://localhost:8004/ocpu/library/finFindR/R/processFinFromHttp/json -F "imageobj=@C:/Users/jathompson/Documents/dolphinTestingdb/jensImgs/test2.jpg"
aka: traceFinFromHttp(imageobj = "yourfile1.jpg")

Processes an image(cimg) containing a fin. 
First the image undergoes cleanup through a variety of filters and glare removal via
\code{constrainSizeFinImage} and \code{fillGlare}
These processes help enhance edge clarity.
The trailing edge is highlighted via neural network. 
The image is then cropped down to the trailing edge for efficiency purposes.
The canny edges are then extracted from the crop and passed to 
\code{traceFromCannyEdges}
which isolates coordinates for the trailing edge. These coordinates are then passed to
\code{extractAnnulus}
which collects image data used for identification.
Both the coordinates and the image annulus are then returned.
}

