% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{msum}
\alias{msum}
\title{msum}
\usage{
msum(x, n = 5, sides = 1)
}
\arguments{
\item{n}{width of kernel}

\item{sides}{should the moving average be calculated in middle, or starting from one side}
}
\value{
smoothed sequence
}
\description{
helper function for smoothing a sequence
}

