% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{traceFromCannyEdges}
\alias{traceFromCannyEdges}
\title{traceFromCannyEdges}
\usage{
traceFromCannyEdges(pathMap, startPoint, endPoint, prox)
}
\arguments{
\item{pathMap}{matrix of type numeric. Used for weighted astar path finding}

\item{startPoint}{vector of type numeric indicating the x and y position for initializing trace}

\item{endPoint}{value of type numeric indicating the x and y position for terminating trace}

\item{prox}{value of type numeric for setting degree of wiggle room for trace termination}
}
\value{
Value of type dataframe containing plotpath coordinates
}
\details{
Isolates trailing edge of a fin given a canny edge image(cimg) and trace constraints.
Starting with a weighted canny edge matrix and start/stop points, an optimal path traversal is found via
\code{findPath}
After some basic cleanup, the resulting pixel coordinates for the optimal edge trace are returned.
}

