% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{traceFromImage}
\alias{traceFromImage}
\title{traceFromImage}
\usage{
traceFromImage(fin, startStopCoords = NULL, pathNet = NULL)
}
\arguments{
\item{fin}{Value of type cimg. Load the image via load.image("directory/finImage.JPG")}

\item{startStopCoords}{list of 3 coordinates: leadingEnd, startPoint, trailingEnd. If NULL, these points are estimated}

\item{pathNet}{mxnet model for isolating trailing edge}
}
\value{
Value of type list containing  a dataframe of coordinates and a cimg of simplified features
}
\details{
Processes an image(cimg) containing a fin. 
First the image undergoes cleanup through a variety of filters and glare removal via
\code{constrainSizeFinImage} and \code{fillGlare}
These processes help enhance edge clarity.
The trailing edge is highlighted via neural network. 
The image is then cropped down to the trailing edge for efficiency purposes.
The canny edges are then extracted from the crop and passed to 
\code{traceFromCannyEdges}
which isolates coordinates for the trailing edge. These coordinates are then passed to
\code{extractAnnulus}
which collects image data used for identification.
Both the coordinates and the image annulus are then returned.
}
