% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{traceToHash}
\alias{traceToHash}
\title{traceToHash}
\usage{
traceToHash(traceData, mxnetModel = NULL)
}
\value{
Value of type list containing:
"coordinates" a dataframe of coordinates
"annulus" a 3 channel image of isolated features
}
\description{
Function which takes the output of \code{traceFromImage} and returns objects used for matching
}

