% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{distanceToRefParallel}
\alias{distanceToRefParallel}
\title{distanceToRefParallel}
\usage{
distanceToRefParallel(queryHashData, referenceHashData, batchSize = 500,
  returnLimit = min(50, length(referenceHashData)),
  counterEnvir = new.env(), displayProgressInShiny = F)
}
\arguments{
\item{queryHashData}{matrix (or list) containing the hashes for matching}

\item{referenceHashData}{matrix (or list) containing a reference catalogue of hashes}

\item{batchSize}{int denoting the number of query instances to process at once}

\item{counterEnvir}{r enfironment object to hold a progress counter for display purposes}

\item{displayProgressInShiny}{bool denoting if function is called inside an rshiny instance}
}
\value{
list of two dataframes: 
"distances" for all query to reference pairs
"sortingIndex" which denote the order from best to worst match in the reference catalogue
}
\description{
Function performing batched matching between a query catalogue to a reference catalogue
To call from opencpu:
curl https://cloud.opencpu.org/ocpu/library/finFindR/R/traceFinFromHttp \
-H 'Content-Type: application/json' \
-d '[
[13.4,-11.6,15,2.4,0.4,12.3,7.7,11.7,-11.3,-0.9,3.5,-1.6,-8.1,-8.8,-1.8,-10.1,-4.8,-16.6,2.1,-4.9,11.7,-16.1,0.9,-4.2,11,3.7,14.8,-1.5,-10.8,11,10.7,-3.3], 
[9.9,-9.7,10.2,-10.9,27.8,-15.2,10,-5.2,7.1,-6.3,3,-0.3,5.4,8.9,13.9,-3.7,3.1,1.9,9.4,-10.1,7.7,1.4,-5.8,-12,12.2,-16.1,-4.5,-5.6,5.3,0,-21.7,2.3]
]
}

